transcriber/requirements.txt
redis==5.0.1
openai-whisper==20231117
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
